refactor: 뷰모델 생성자 기본값을 제공하여 테스트 통과하도록 수정
